Refactored recipe service to remove utility methods.
Moved utility methods out of the recipe service and into utility objects
in appropriate files. Thus, utility methods related to filtering recipes
were placed in a `filterUtil` object in the `recipe-filter.ts` file.
Similarly, methods like `addIngredientTo` were placed in a `recipeUtil`
object exported from `recipe.ts`.
