+++ src/app/recipes/recipe.service.ts
@@ -9,7 +9,6 @@ import * as firebase                    from 'firebase/app';
 import { UserService }                  from "../user/user.service";
 import { Recipe }                       from "./recipe";
 import { RecipeGist }                   from "./recipe-gist";
-import { RecipeFilter, applyFilter }    from "./recipe-filter";
 
 @Injectable()
 export class RecipeService {
@@ -25,20 +24,6 @@ export class RecipeService {
   getCurrentUserRecipeGists(): Observable<RecipeGist[]> {
     return this.userService.userIdSubject
       .switchMap(id => id ? this.getUserRecipeGists(id) : []);
-  }
-
-  // Filter the recipe array of a passed observable
-  filter(
-    filter: RecipeFilter, recipeGists: Observable<RecipeGist[]>
-  ): Observable<RecipeGist[]> {
-    return recipeGists.map(gists => applyFilter(filter, gists));
-  }
-
-  // Filter the recipe array of a passed observable by tags
-  filterByTags(
-    tags: string[], gists: Observable<RecipeGist[]>
-  ): Observable<RecipeGist[]> {
-    return this.filter({ hasTags: tags }, gists);
   }
 
   getRecipe(id: string): Observable<Recipe> {
@@ -101,56 +86,5 @@ export class RecipeService {
 
     await recUpdate;
     await gistUpdate;
-  }
-
-
-  // Add an ingredient to a recipe
-  // Here because the recipe class works more as an interface than a class
-  addIngredientTo(recipe: Recipe, ingredient: string): void {
-    // Ensure there is a recipe and an ingredient
-    if (recipe && ingredient)
-      // If the ingredients field does exist
-      if (recipe.ingredients)
-        // Push the ingredient on
-        recipe.ingredients.push(ingredient);
-      // If the ingredients field doesn't exist
-      else
-        // Set it to an array with the ingredient already stored
-        recipe.ingredients = [ingredient];
-  }
-
-  // Delete an ingredient from a recipe
-  // Here because the recipe class works more as an interface than a class
-  deleteIngredientFrom(recipe: Recipe, ingredient: string): void {
-    // Ensure there is a recipe and an ingredient
-    if (recipe && ingredient)
-      // Ensure the ingredients field exists
-      if (recipe.ingredients) {
-        // Find the index of the ingredient
-        let i = recipe.ingredients.indexOf(ingredient);
-        // If the ingredient exists (i.e., the index isn't -1)
-        if (i !== -1)
-          // Delete the ingredient
-          recipe.ingredients.splice(i, 1);
-      }
-  }
-
-
-  // Add a tag to a recipe
-  addTagTo(recipe: Recipe, tag: string): void {
-    if (recipe && tag)
-      if (recipe.tags)
-        recipe.tags.push(tag);
-      else
-        recipe.tags = [tag];
-  }
-
-  // Remove a tag from a recipe
-  deleteTagFrom(recipe: Recipe, tag: string): void {
-    if (recipe && tag && recipe.tags) {
-      let i = recipe.tags.indexOf(tag);
-
-      if (i !== -1) recipe.tags.splice(i, 1);
-    }
   }
 }
